+++ log.cpp
@@ -3,7 +3,6 @@
 #include "embark.h"
 #include "plan.h"
 #include "population.h"
-#include "plan_setup.h"
 
 #include "modules/Gui.h"
 #include "modules/Translation.h"
@@ -312,7 +311,7 @@ void AI::event(const std::string & name, const Json::Value & payload)
 
 std::string AI::status()
 {
-    if (events.has_exclusive<EmbarkExclusive>() || events.has_exclusive<PlanSetup>())
+    if (is_embarking())
     {
         return "(embarking)";
     }
@@ -346,7 +345,7 @@ static void report_section(std::ostringstream & out, const std::string & name, M
 
 std::string AI::report(bool html)
 {
-    if (events.has_exclusive<EmbarkExclusive>() || events.has_exclusive<PlanSetup>())
+    if (is_embarking())
     {
         if (html)
         {