+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `allow_pause` config setting (defaults to true, set to false for old behavior).
 - df-ai will now act as a client of dfplex if dfplex is enabled.
 - Fixed a long-standing bug that could cause the game to completely freeze if lockstep was disabled under certain conditions.
+- The weblegends pages df-ai adds are now rendered in the faux Wikipedia style of the rest of weblegends.
 
 # 0.47.04-beta1